+++ src/app/interfaces/index.ts
@@ -1,1 +1,1 @@
-export * from './elementConfig'
+export * from './customHtmlElementModel'
